Moved shared folder into the app folder and updated all of the filepath references. Also removed the changeDetection strategy since it's not really being used yet and was causing some problems.
