+++ src/app/apiKeys/apiKeys.ts
@@ -3,7 +3,7 @@
 export { Component } from '@angular/core';
 export { IAPIKeys } from './apiKeys.interface';
 export { APIKeysService } from './apiKeys.service';
-export { ButtonComponent } from '../../shared/button/button.component';
-export { FilterComponent } from '../../shared/filter/filter.component';
-export { BreadcrumbComponent } from '../../shared/breadcrumbs/breadcrumbs.component';
-export { ButtonSaveComponent } from '../../shared/buttonSave/buttonSave.component';
+export { ButtonComponent } from '../shared/button/button.component';
+export { FilterComponent } from '../shared/filter/filter.component';
+export { BreadcrumbComponent } from '../shared/breadcrumbs/breadcrumbs.component';
+export { ButtonSaveComponent } from '../shared/buttonSave/buttonSave.component';
