+++ src/app/deployment/deployment.ts
@@ -1,4 +1,4 @@
 //barrel file for deployment settings
 
 export { Component } from '@angular/core';
-export { BreadcrumbComponent } from '../../shared/breadcrumbs/breadcrumbs.component';
+export { BreadcrumbComponent } from '../shared/breadcrumbs/breadcrumbs.component';
