+++ src/app/general/general.ts
@@ -1,10 +1,10 @@
 //barrel file for general settings
 
 export { Component } from '@angular/core';
-export { BreadcrumbComponent } from '../../shared/breadcrumbs/breadcrumbs.component';
+export { BreadcrumbComponent } from '../shared/breadcrumbs/breadcrumbs.component';
 export { NgForm } from '@angular/forms';
 export { GeneralSettingsService } from './general.service';
 export { IPointOfContact, IBusinessAddress, BusinessAddress } from './general.interface';
 export { MD_INPUT_DIRECTIVES } from '@angular2-material/input/input';
 export { MD_CARD_DIRECTIVES } from '@angular2-material/card/card';
-export { ButtonSaveComponent } from '../../shared/buttonSave/buttonSave.component';
+export { ButtonSaveComponent } from '../shared/buttonSave/buttonSave.component';
