+++ src/app/lights/lights.ts
@@ -9,7 +9,7 @@ export { LightsService } from './lights.service';
 export { LightsReducer } from './lights.reducer';
 export { LightsList } from './components/lights.list';
 export { LightsDetail } from './components/lights.detail';
-export { ButtonComponent } from '../../shared/button/button.component';
-export { FilterComponent } from '../../shared/filter/filter.component';
-export { BreadcrumbComponent } from '../../shared/breadcrumbs/breadcrumbs.component';
+export { ButtonComponent } from '../shared/button/button.component';
+export { FilterComponent } from '../shared/filter/filter.component';
+export { BreadcrumbComponent } from '../shared/breadcrumbs/breadcrumbs.component';
 export { Observable } from 'rxjs/Observable';
