+++ src/app/settings/settings.ts
@@ -1,4 +1,4 @@
 //barrel file for settings
 
 export { Component } from '@angular/core';
-export { SidebarComponent } from '../../shared/sidebar/sidebar.component';
+export { SidebarComponent } from '../shared/sidebar/sidebar.component';
