+++ src/app/shared/breadcrumbs/breadcrumbs.component.less
@@ -1,4 +1,4 @@
-@import "../../css/colors.less";
+@import "../../../css/colors.less";
 
 .breadcrumb{
     @color: @graytext;
