+++ src/app/shared/button/button.component.less
@@ -1,4 +1,4 @@
-@import "../../css/colors.less";
+@import "../../../css/colors.less";
 
 .md-primary[md-raised-button] {
     background-color: @highlightblue !important;
