+++ src/app/shared/buttonAdd/buttonAdd.component.less
@@ -1,4 +1,4 @@
-@import "../../css/colors.less";
+@import "../../../css/colors.less";
 
 button {
     background-color: @mainblue;
