+++ src/app/shared/buttonSave/buttonSave.component.less
@@ -1,4 +1,4 @@
-@import "../../css/colors.less";
+@import "../../../css/colors.less";
 
 .md-accent[md-fab]{
     background-color: @mainblue !important;
