+++ src/app/shared/filter/filter.component.less
@@ -1,4 +1,4 @@
-@import "../../css/colors.less";
+@import "../../../css/colors.less";
 
 .filter {
     display: inline-block;
