+++ src/app/shared/map/map.component.spec.ts
