+++ src/app/shared/sidebar/README.md
@@ -3,7 +3,7 @@
 #### Import the SidebarComponent and add it to the directives.
 
 ```typescript
-import { SidebarComponent } from '../../shared/sidebar/sidebar.component';
+import { SidebarComponent } from '../shared/sidebar/sidebar.component';
 ```
 
 #### Create an array of objects with name and path properties.
