+++ src/app/shared/sidebar/sidebar.component.less
@@ -1,4 +1,4 @@
-@import "../../css/colors.less";
+@import "../../../css/colors.less";
 
 /* MIXINS */
