+++ src/app/sites-all/containers/sites-all.container.ts
@@ -1,13 +1,12 @@
 import { Component, Input, Output, EventEmitter, ChangeDetectionStrategy } from '@angular/core';
 import { ISite, AllSitesService, ButtonComponent, FilterComponent, BreadcrumbComponent } from '../sites-all';
-import { MapComponent } from '../../../shared/map/map.component';
 
 //import less
 import './sites-all.container.less';
 
 @Component({
     selector: 'app-sites-all',
-    directives: [ButtonComponent, FilterComponent, BreadcrumbComponent, MapComponent],
+    directives: [ButtonComponent, FilterComponent, BreadcrumbComponent],
     templateUrl: '../src/app/sites-all/containers/sites-all.container.html',
     providers: [AllSitesService]
 })
