+++ src/app/sites-all/sites-all.ts
@@ -4,6 +4,6 @@ export { Component, Input, Output, EventEmitter, ChangeDetectionStrategy } from 
 export { ISite } from './sites-all.interface';
 export { AppStore } from '../app.store';
 export { AllSitesService } from './sites-all.service';
-export { ButtonComponent } from '../../shared/button/button.component';
-export { FilterComponent } from '../../shared/filter/filter.component';
-export { BreadcrumbComponent } from '../../shared/breadcrumbs/breadcrumbs.component';
+export { ButtonComponent } from '../shared/button/button.component';
+export { FilterComponent } from '../shared/filter/filter.component';
+export { BreadcrumbComponent } from '../shared/breadcrumbs/breadcrumbs.component';
