+++ src/app/sites-groups/sites-groups.ts
@@ -3,6 +3,6 @@
 export { Component } from '@angular/core';
 export { ISiteGroups } from './sites-groups.interface';
 export { SiteGroupsService } from './sites-groups.service';
-export { ButtonComponent } from '../../shared/button/button.component';
-export { FilterComponent } from '../../shared/filter/filter.component';
-export { BreadcrumbComponent } from '../../shared/breadcrumbs/breadcrumbs.component';
+export { ButtonComponent } from '../shared/button/button.component';
+export { FilterComponent } from '../shared/filter/filter.component';
+export { BreadcrumbComponent } from '../shared/breadcrumbs/breadcrumbs.component';
