+++ src/app/sites/sites.ts
@@ -2,7 +2,7 @@
 
 export { Component } from '@angular/core';
 export { ROUTER_DIRECTIVES } from '@angular/router';
-export { ButtonAddComponent } from '../../shared/buttonAdd/buttonAdd.component'; 
+export { ButtonAddComponent } from '../shared/buttonAdd/buttonAdd.component'; 
 export { SitesAllContainer } from '../sites-all/containers/sites-all.container';
 export { SitesGroupsComponent } from '../sites-groups/sites-groups.component';
-export { MapComponent } from '../../shared/map/map.component';
+export { MapComponent } from '../shared/map/map.component';
