+++ src/app/users/users.ts
@@ -1,5 +1,5 @@
 //barrel file for users
 
 export { Component } from '@angular/core';
-export { BreadcrumbComponent } from '../../shared/breadcrumbs/breadcrumbs.component';
+export { BreadcrumbComponent } from '../shared/breadcrumbs/breadcrumbs.component';
 export { AuthenticationComponent } from '../authentication/authentication.component';
